requirements.txt
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
blinker==1.4
click==7.1.2
colorama==0.4.3
coverage==5.1
docopt==0.6.2
flake8==3.8.1
Flask==1.1.2
Flask-Mail==0.9.1
flask-marshmallow==0.12.0
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.6.0
marshmallow-sqlalchemy==0.23.0
mccabe==0.6.1
more-itertools==8.2.0
nose==1.3.7
nose2==0.9.2
packaging==20.3
pathtools==0.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytest==5.4.2
pytest-testmon==1.0.2
python-dotenv==0.13.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.16
toml==0.10.0
watchdog==0.10.2
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
